+++ de.bund.bfr.knime.fsklab.nodes.test/build.properties
@@ -1,5 +1,4 @@
-source.. = src/,\
-           workflows/
+source.. = src/java/
 output.. = bin/
 bin.includes = META-INF/,\
                .
