+++ de.bund.bfr.knime.fsklab.nodes.test/src/java/de/bund/bfr/knime/fsklab/nodes/WorkflowsTest.java
@@ -1,0 +1,35 @@
+package de.bund.bfr.knime.fsklab.nodes;
+
+import java.io.File;
+
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ErrorCollector;
+import org.knime.testing.core.TestrunConfiguration;
+
+import nl.esciencecenter.e3dchem.knime.testing.TestFlowRunner;
+
+public class WorkflowsTest {
+
+	@Rule
+	public ErrorCollector collector = new ErrorCollector();
+	
+	private TestFlowRunner runner;
+	
+	@Before
+	public void setUp() {
+		TestrunConfiguration runConfiguration = new TestrunConfiguration();
+		runConfiguration.setTestDialogs(false);
+		runConfiguration.setReportDeprecatedNodes(true);
+		runConfiguration.setCheckMemoryLeaks(false);
+		runConfiguration.setLoadSaveLoad(false);
+		runner = new TestFlowRunner(collector, runConfiguration);
+	}
+	
+	@Test
+	public void testChickenModels() throws Exception {
+		File workflowDir = new File("src/workflows/InitializeParentsFlocks");
+		runner.runTestWorkflow(workflowDir);
+	}
+}
