+++ de.bund.bfr.knime.fsklab.nodes.test/src/workflows/InitializeParentsFlocks/FSK Creator (#1)/drop/viz.R
@@ -1,0 +1,1 @@
+hist(result, breaks=50, main="PREVALENCE WITHIN A FLOCK IN PARENT GENERATION", xlab="Prevalence", col="32")
