+++ de.bund.bfr.knime.fsklab.nodes.test/src/workflows/InitializeParentsFlocks/FSK Runner (#3)/settings.xml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<config xmlns="http://www.knime.org/2008/09/XMLConfig" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.knime.org/2008/09/XMLConfig http://www.knime.org/XMLConfig_2008_09.xsd" key="settings.xml">
+<entry key="node_file" type="xstring" value="settings.xml"/>
+<config key="flow_stack"/>
+<config key="internal_node_subsettings">
+<entry key="memory_policy" type="xstring" value="CacheSmallInMemory"/>
+</config>
+<config key="model">
+<entry key="width" type="xint" value="640"/>
+<entry key="height" type="xint" value="640"/>
+<entry key="resolution" type="xstring" value="NA"/>
+<entry key="textPointSize" type="xint" value="12"/>
+<entry key="simulation" type="xstring" value="defaultSimulation"/>
+</config>
+<config key="nodeAnnotation">
+<entry key="text" type="xstring" value=""/>
+<entry key="bgcolor" type="xint" value="16777215"/>
+<entry key="x-coordinate" type="xint" value="247"/>
+<entry key="y-coordinate" type="xint" value="179"/>
+<entry key="width" type="xint" value="106"/>
+<entry key="height" type="xint" value="15"/>
+<entry key="alignment" type="xstring" value="CENTER"/>
+<entry key="borderSize" type="xint" value="0"/>
+<entry key="borderColor" type="xint" value="16777215"/>
+<entry key="defFontSize" type="xint" value="9"/>
+<entry key="annotation-version" type="xint" value="20151123"/>
+<config key="styles"/>
+</config>
+<entry key="customDescription" type="xstring" isnull="true" value=""/>
+<entry key="state" type="xstring" value="CONFIGURED"/>
+<entry key="factory" type="xstring" value="de.bund.bfr.knime.fsklab.nodes.RunnerNodeFactory"/>
+<entry key="node-name" type="xstring" value="FSK Runner"/>
+<entry key="node-bundle-name" type="xstring" value="BfR FSK-Lab extension nodes"/>
+<entry key="node-bundle-symbolic-name" type="xstring" value="de.bund.bfr.knime.fsklab.nodes"/>
+<entry key="node-bundle-vendor" type="xstring" value="German Federal Institute for Risk Assessment (BfR)"/>
+<entry key="node-bundle-version" type="xstring" value="1.0.0.201812180923"/>
+<entry key="node-feature-name" type="xstring" value="BfR FSK-Lab"/>
+<entry key="node-feature-symbolic-name" type="xstring" value="de.bund.bfr.knime.fsklab.feature.feature.group"/>
+<entry key="node-feature-vendor" type="xstring" value="German Federal Institute for Risk Assessment (BfR)"/>
+<entry key="node-feature-version" type="xstring" value="1.4.2.201812180923"/>
+<config key="factory_settings"/>
+<entry key="name" type="xstring" value="FSK Runner"/>
+<entry key="hasContent" type="xboolean" value="false"/>
+<entry key="isInactive" type="xboolean" value="false"/>
+<config key="ports">
+<config key="port_1">
+<entry key="index" type="xint" value="1"/>
+<entry key="port_dir_location" type="xstring" isnull="true" value=""/>
+</config>
+<config key="port_2">
+<entry key="index" type="xint" value="2"/>
+<entry key="port_dir_location" type="xstring" isnull="true" value=""/>
+</config>
+</config>
+<config key="filestores">
+<entry key="file_store_location" type="xstring" isnull="true" value=""/>
+<entry key="file_store_id" type="xstring" isnull="true" value=""/>
+</config>
+</config>
