+++ de.bund.bfr.knime.fsklab.nodes.test/src/workflows/InitializeParentsFlocks/Testflow Configuration (#2)/settings.xml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<config xmlns="http://www.knime.org/2008/09/XMLConfig" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.knime.org/2008/09/XMLConfig http://www.knime.org/XMLConfig_2008_09.xsd" key="settings.xml">
+<entry key="node_file" type="xstring" value="settings.xml"/>
+<config key="flow_stack"/>
+<config key="internal_node_subsettings">
+<entry key="memory_policy" type="xstring" value="CacheSmallInMemory"/>
+</config>
+<config key="model">
+<entry key="owner" type="xstring" value="miguel.de-alba-aparicio@bfr.bund.de"/>
+<config key="requiredLogErrors">
+<entry key="array-size" type="xint" value="0"/>
+</config>
+<config key="requiredLogWarnings">
+<entry key="array-size" type="xint" value="0"/>
+</config>
+<config key="requiredLogInfos">
+<entry key="array-size" type="xint" value="0"/>
+</config>
+<config key="optionalLogMessages">
+<entry key="array-size" type="xint" value="0"/>
+</config>
+<config key="failingNodes">
+<entry key="array-size" type="xint" value="0"/>
+</config>
+<config key="requiredNodeWarnings">
+<entry key="count" type="xint" value="0"/>
+</config>
+<config key="requiredNodeErrors">
+<entry key="count" type="xint" value="0"/>
+</config>
+<entry key="timeout" type="xint" value="300"/>
+<entry key="maxHilitedRows" type="xint" value="2500"/>
+<config key="usedJanitors">
+<entry key="array-size" type="xint" value="0"/>
+</config>
+<entry key="streamingTest" type="xboolean" value="false"/>
+<entry key="requiredLoadVersion" type="xstring" value="&lt;future&gt;"/>
+</config>
+<config key="nodeAnnotation">
+<entry key="text" type="xstring" value=""/>
+<entry key="bgcolor" type="xint" value="16777215"/>
+<entry key="x-coordinate" type="xint" value="27"/>
+<entry key="y-coordinate" type="xint" value="179"/>
+<entry key="width" type="xint" value="106"/>
+<entry key="height" type="xint" value="15"/>
+<entry key="alignment" type="xstring" value="CENTER"/>
+<entry key="borderSize" type="xint" value="0"/>
+<entry key="borderColor" type="xint" value="16777215"/>
+<entry key="defFontSize" type="xint" value="9"/>
+<entry key="annotation-version" type="xint" value="20151123"/>
+<config key="styles"/>
+</config>
+<entry key="customDescription" type="xstring" isnull="true" value=""/>
+<entry key="state" type="xstring" value="CONFIGURED"/>
+<entry key="factory" type="xstring" value="org.knime.testing.node.config.TestConfigNodeFactory"/>
+<entry key="node-name" type="xstring" value="Testflow Configuration"/>
+<entry key="node-bundle-name" type="xstring" value="Test-Node extension for KNIME Workbench"/>
+<entry key="node-bundle-symbolic-name" type="xstring" value="org.knime.testing"/>
+<entry key="node-bundle-vendor" type="xstring" value="KNIME AG, Zurich, Switzerland"/>
+<entry key="node-bundle-version" type="xstring" value="3.5.0.v201712051533"/>
+<entry key="node-feature-name" type="xstring" value="KNIME Testing Framework"/>
+<entry key="node-feature-symbolic-name" type="xstring" value="org.knime.features.testingapplication.feature.group"/>
+<entry key="node-feature-vendor" type="xstring" value="KNIME AG, Zurich, Switzerland"/>
+<entry key="node-feature-version" type="xstring" value="3.5.2.v201802051355"/>
+<config key="factory_settings"/>
+<entry key="name" type="xstring" value="Testflow Configuration"/>
+<entry key="hasContent" type="xboolean" value="false"/>
+<entry key="isInactive" type="xboolean" value="false"/>
+<config key="ports"/>
+<config key="filestores">
+<entry key="file_store_location" type="xstring" isnull="true" value=""/>
+<entry key="file_store_id" type="xstring" isnull="true" value=""/>
+</config>
+</config>
