--- de.bund.bfr.knime.fsklab.nodes.test/workflows/InitializeParentsFlocks/.project
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<projectDescription>
-	<name>InitializeParentsFlocks</name>
-	<comment></comment>
-	<projects>
-	</projects>
-	<buildSpec>
-	</buildSpec>
-	<natures>
-		<nature>org.knime.workbench.ui.KNIMEProjectNature</nature>
-	</natures>
-</projectDescription>
