--- de.bund.bfr.knime.fsklab.nodes.test/workflows/InitializeParentsFlocks/.savedWithData
@@ -1,4 +1,0 @@
-Do not delete this file!
-This file serves to indicate that the workflow was written as part of the usual save routine (not exported).
-
-Workflow was last saved by user de on Tue Dec 18 15:01:29 CET 2018
